LFD tests reported
All secondary school pupils should receive 2 on-site LFD tests, 3 to 5 days apart, on their return in the autumn term. Students as well as staff should then continue to test twice weekly at home  until the end of September, when this will be reviewed. For more information on testing in  schools the different phases of education see the schools COVID-19 operational guidance